Change cloudinary fetch from App.js to CameraPage.js and update the submit-function
