+++ src/App.js
@@ -1,17 +1,11 @@
 import { useState } from 'react'
-import axios from 'axios'
 import background from '../src/assets/viary-bg2.jpg'
 import Header from '../src/components/Header'
 import CameraPage from './pages/CameraPage'
-import MemoriesPage from './pages/MemoriesPage'
 import DetailMemoryPage from './pages/DetailMemoryPage'
-
-const CLOUDNAME = process.env.REACT_APP_CLOUDINARY_CLOUDNAME
-const PRESET = process.env.REACT_APP_CLOUDINARY_PRESET
+import MemoriesPage from './pages/MemoriesPage'
 
 export default function App() {
-  const [image, setImage] = useState('')
-  const [currentMemory, setCurrentMemory] = useState({})
   const [memories, setMemories] = useState([])
   const [currentPage, setCurrentPage] = useState('camera')
   const [detailImage, setDetailImage] = useState(null)
@@ -32,12 +26,7 @@ export default function App() {
     >
       <Header>Viary</Header>
       {currentPage === 'camera' && (
-        <CameraPage
-          image={image}
-          upload={upload}
-          onNavigate={showMemoriesPage}
-          onSubmit={handleSubmit}
-        />
+        <CameraPage onNavigate={showMemoriesPage} onSubmit={handleSubmit} />
       )}
       {currentPage === 'memories' && (
         <MemoriesPage
@@ -56,31 +45,8 @@ export default function App() {
     </div>
   )
 
-  function upload(event) {
-    const url = `https://api.cloudinary.com/v1_1/${CLOUDNAME}/upload`
-
-    const formData = new FormData()
-    formData.append('file', event.target.files[0])
-    formData.append('upload_preset', PRESET)
-
-    axios
-      .post(url, formData, {
-        headers: {
-          'Content-type': 'multipart/form-data',
-        },
-      })
-      .then(onImageSave)
-      .catch(err => console.error(err))
-  }
-
-  function onImageSave(response) {
-    setImage(response.data.url)
-    setCurrentMemory({ image: response.data.url, ...currentMemory })
-  }
-
   function handleSubmit(newMemory) {
-    setCurrentMemory({ ...newMemory, ...currentMemory })
-    setMemories([currentMemory, ...memories])
+    setMemories([newMemory, ...memories])
     setCurrentPage('memories')
   }
